pyproject.toml
[tool.poetry]
name = "mytardis-rocrate-builder"
version = "0.1.0"
description = ""
authors = ["James Love <james.love@auckland.ac.nz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
slugify = "^0.0.1"
ro-crate-py = {git = "https://github.com/UoA-eResearch/ro-crate-py.git",  branch = "encrypted-metadata"}
bagit = "^1.8.1"
types-python-slugify = "^8.0.2.20240310"
mypy = "^1.10.0"
pylint = "^3.2.2"
pydantic = "^2.7.1"
black = "^24.4.2"
uuid = "^1.30"
validators = "^0.28.3"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disable_error_code = "import-untyped"
files = [
       "**/*.py"
	  ]
exclude = [
       "^tests/.*$"
       ]
strict = true
ignore_missing_imports = true

[tool.isort]
profile = "black"


